// src/common/enums/customError.enum.ts
export enum CustomError {
  UserNotExist = "User don't exist",
  RoomNotExist = "Room don'texist",
  AlreadyExist = 'Already exist',
  PermissionError = 'Permission error',
  WrongRequestBody = 'Wrong request body',
  WrongLoginOrPassword = 'Wrong login or password',
  UsernameIncompatibleWithPattern = 'Username Incompatible With Pattern',
}